+++ lib/ar_lazy_preload/context.rb
@@ -9,13 +9,17 @@ module ArLazyPreload
   # Calling #preload_association method will cause loading of ALL associated objects for EACH
   # ecord when requested association is found in the association tree.
   class Context
-    attr_reader :model, :records, :association_tree
+    # Initiates lazy preload context for given records
+    def self.register(records:, association_tree:)
+      return if records.empty? || association_tree.empty?
+      ArLazyPreload::Context.new(records: records, association_tree: association_tree)
+    end
 
-    # :model - ActiveRecord class which records belong to
+    attr_reader :records, :association_tree
+
     # :records - array of ActiveRecord instances
     # :association_tree - list of symbols or hashes representing a tree of preloadable associations
-    def initialize(model:, records:, association_tree:)
-      @model = model
+    def initialize(records:, association_tree:)
       @records = records.compact
       @association_tree = association_tree
 
@@ -26,13 +30,8 @@ module ArLazyPreload
     # objects in the context it if needed.
     def try_preload_lazily(association_name)
       return unless association_needs_preload?(association_name)
-
       preloader.preload(records, association_name)
-
-      AssociatedContextBuilder.new(
-        parent_context: self,
-        association_name: association_name
-      ).perform
+      AssociatedContextBuilder.prepare(parent_context: self, association_name: association_name)
     end
 
     private