+++ spec/ar_lazy_preload_spec.rb
@@ -156,13 +156,22 @@ describe ArLazyPreload do
   describe "polymorphic" do
     include_examples "check initial loading"
 
-    subject { Vote.lazy_preload(:voteable) }
+    subject { Vote.lazy_preload(voteable: :user) }
 
     # SELECT "votes".* FROM "votes"
     # SELECT "posts".* FROM "posts" WHERE "posts"."id" IN (...)
     # SELECT "comments".* FROM "comments" WHERE "comments"."id" IN (...)
     it "loads lazy_preloaded association" do
       expect { subject.map { |vote| vote.voteable.id } }.to make_database_queries(count: 3)
+    end
+
+    # SELECT "votes".* FROM "votes"
+    # SELECT "posts".* FROM "posts" WHERE "posts"."id" IN (...)
+    # SELECT "comments".* FROM "comments" WHERE "comments"."id" IN (...)
+    # SELECT "users".* FROM "users" WHERE "users"."id" IN (...) - for posts
+    # SELECT "users".* FROM "users" WHERE "users"."id" IN (...) - for comments
+    it "loads embedded lazy_preloaded association" do
+      expect { subject.map { |vote| vote.voteable.user.id } }.to make_database_queries(count: 5)
     end
   end