+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- [PR [#1](https://github.com/DmitryTsepelev/ar_lazy_preload/pull/1)] Cleanup code, fix initiating child context for preload associations ([@DmitryTsepelev][])
+
 ## 0.1.0 (2018-08-01)
 
 - Initial version. ([@DmitryTsepelev][])